+++ src/main/java/http/request/HttpRequestBody.java
@@ -1,4 +1,4 @@
-package http;
+package http.request;
 
 public class HttpRequestBody {
     private final String content;
