+++ src/main/java/http/request/HttpRequestHeader.java
@@ -1,6 +1,5 @@
-package http;
+package http.request;
 
-import java.util.HashMap;
 import java.util.Map;
 
 public class HttpRequestHeader {
